+++ opentelemetry-api/src/opentelemetry/trace/__init__.py
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+# pylint: disable=no-name-in-module
 
 """
 The OpenTelemetry tracing API describes the classes used to generate
@@ -84,7 +85,9 @@ from typing import Iterator, Optional, Sequence, cast
 from opentelemetry import context as context_api
 from opentelemetry.attributes import BoundedAttributes  # type: ignore
 from opentelemetry.context.context import Context
-from opentelemetry.environment_variables import OTEL_PYTHON_TRACER_PROVIDER
+from opentelemetry.environment_variables import (
+    OTEL_PYTHON_TRACER_PROVIDER,  # type: ignore
+)
 from opentelemetry.trace.propagation import (
     _SPAN_KEY,
     get_current_span,
@@ -486,14 +489,14 @@ def get_tracer_provider() -> TracerProvider:
     if _TRACER_PROVIDER is None:
         # if a global tracer provider has not been set either via code or env
         # vars, return a proxy tracer provider
-        if OTEL_PYTHON_TRACER_PROVIDER not in os.environ:
+        if OTEL_PYTHON_TRACER_PROVIDER not in os.environ:  # type: ignore
             if not _PROXY_TRACER_PROVIDER:
                 _PROXY_TRACER_PROVIDER = ProxyTracerProvider()
             return _PROXY_TRACER_PROVIDER
 
         _TRACER_PROVIDER = cast(  # type: ignore
             "TracerProvider",
-            _load_provider(OTEL_PYTHON_TRACER_PROVIDER, "tracer_provider"),
+            _load_provider(OTEL_PYTHON_TRACER_PROVIDER, "tracer_provider"),  # type: ignore
         )
     return _TRACER_PROVIDER