+++ opentelemetry-distro/src/opentelemetry/distro/__init__.py
@@ -11,7 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
+# pylint: disable=no-name-in-module
+
 import os
 from logging import getLogger
 from os import environ