+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/open-telemetry/opentelemetry-python/compare/v1.3.0-0.22b0...HEAD)
 - `opentelemetry-semantic-conventions` Generate semconv constants update for OTel Spec 1.5.0
   ([#1946](https://github.com/open-telemetry/opentelemetry-python/pull/1946))
+- Make all environment variables accessible from `opentelemetry.environment_variables`.
+  ([#1946](https://github.com/open-telemetry/opentelemetry-python/pull/1946))
 
 ### Added
 - Moved `opentelemetry-instrumentation` to core repository.